+++ lib/ggity/scale/linetype_discrete.ex
@@ -59,8 +59,7 @@ defmodule GGity.Scale.Linetype.Discrete do
         "#{label}",
         x: "0",
         y: "-5",
-        font_size: "9",
-        fill: "black",
+        class: "gg-text gg-legend-title",
         text_anchor: "left"
       ),
       Stream.with_index(levels)
@@ -73,19 +72,16 @@ defmodule GGity.Scale.Linetype.Discrete do
       Draw.rect(
         x: "0",
         y: "#{15 * index}",
-        height: "15",
-        width: "15",
-        fill: "#eeeeee",
-        stroke: "#eeeeee",
-        stroke_width: "0.5"
+        height: 15,
+        width: 15,
+        class: "gg-legend-key"
       ),
       draw_key_glyph(scale, level, index, key_glyph),
       Draw.text(
         "#{Labels.format(scale, level)}",
         x: "20",
         y: "#{10 + 15 * index}",
-        font_size: "8",
-        fill: "black",
+        class: "gg-text gg-legend-text",
         text_anchor: "left"
       )
     ]
@@ -113,16 +109,5 @@ defmodule GGity.Scale.Linetype.Discrete do
       1,
       scale.transform.(level)
     )
-  end
-
-  @spec legend_height(Linetype.Discrete.t()) :: non_neg_integer()
-  def legend_height(%Linetype.Discrete{guide: :none}), do: 0
-
-  def legend_height(%Linetype.Discrete{levels: []}), do: 0
-
-  def legend_height(%Linetype.Discrete{levels: [_]}), do: 0
-
-  def legend_height(%Linetype.Discrete{} = scale) do
-    20 + 15 * length(scale.levels)
   end
 end