+++ lib/ggity/scale/size_discrete.ex
@@ -51,8 +51,7 @@ defmodule GGity.Scale.Size.Discrete do
         "#{label}",
         x: "0",
         y: "-5",
-        font_size: "9",
-        fill: "black",
+        class: "gg-text gg-legend-title",
         text_anchor: "left"
       ),
       Stream.with_index(levels)
@@ -65,19 +64,16 @@ defmodule GGity.Scale.Size.Discrete do
       Draw.rect(
         x: "0",
         y: "#{15 * index}",
-        height: "15",
-        width: "15",
-        fill: "#eeeeee",
-        stroke: "#eeeeee",
-        stroke_width: "0.5"
+        height: 15,
+        width: 15,
+        class: "gg-legend-key"
       ),
       draw_key_glyph(scale, level, index, key_glyph),
       Draw.text(
         "#{Labels.format(scale, level)}",
         x: "20",
         y: "#{10 + 15 * index}",
-        font_size: "8",
-        fill: "black",
+        class: "gg-text gg-legend-text",
         text_anchor: "left"
       )
     ]
@@ -102,16 +98,5 @@ defmodule GGity.Scale.Size.Discrete do
       fill: "black",
       fill_opacity: "1"
     )
-  end
-
-  @spec legend_height(Size.Discrete.t()) :: non_neg_integer()
-  def legend_height(%Size.Discrete{guide: :none}), do: 0
-
-  def legend_height(%Size.Discrete{levels: []}), do: 0
-
-  def legend_height(%Size.Discrete{levels: [_]}), do: 0
-
-  def legend_height(%Size.Discrete{} = scale) do
-    20 + 15 * length(scale.levels)
   end
 end